LayoutTests/imported/w3c/web-platform-tests/shadow-dom/focus/ShadowRoot-delegatesFocus-expected.txt


FAIL default delegatesFocus value assert_equals: expected (boolean) false but got (undefined) undefined
FAIL delegatesFocus set to false in init dict assert_equals: expected (boolean) false but got (undefined) undefined
FAIL delegatesFocus set to true in init dict assert_equals: expected (boolean) true but got (undefined) undefined
